Use StateObject instead of ObservedObject.
From the Apple documentation: "SwiftUI might create or recreate a view at any time, so it’s important that initializing a view with a given set of inputs always results in the same view. As a result, it’s unsafe to create an observed object inside a view. Instead, SwiftUI provides the StateObject attribute for this purpose."

https://developer.apple.com/documentation/swiftui/managing-model-data-in-your-app